+++ lizard_progress/views/views.py
@@ -17,6 +17,7 @@ import json
 
 from matplotlib import figure
 from matplotlib.backends.backend_agg import FigureCanvasAgg as FigureCanvas
+
 from django import http
 from django.conf import settings
 from django.contrib import auth
@@ -25,15 +26,16 @@ from django.contrib.auth.decorators import login_required
 from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import PermissionDenied
 from django.core.urlresolvers import reverse
-from django.db import connection
 from django.http import Http404
 from django.http import HttpResponse
 from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404
+<<<<<<< HEAD
 from django.template import RequestContext
 from django.template.loader import render_to_string
+=======
+>>>>>>> master
 from django.utils.functional import cached_property
-from django.utils.text import Truncator
 from django.utils.translation import ugettext as _
 from django.views.generic.base import TemplateView
 from django.views.decorators.csrf import csrf_exempt
@@ -54,13 +56,15 @@ from lizard_progress.email_notifications.models import NotificationSubscription
 from lizard_progress.email_notifications.models import NotificationType
 from lizard_progress.models import Location
 from lizard_progress.models import MeasurementTypeAllowed
-from lizard_progress.models import Measurement
 from lizard_progress.models import Project
 from lizard_progress.models import ReviewProject
 from lizard_progress.models import has_access
 from lizard_progress.models import has_access_reviewproject
 from lizard_progress.models import Activity
+<<<<<<< HEAD
 from lizard_progress.models import UserProfile
+=======
+>>>>>>> master
 from lizard_progress.models import AvailableMeasurementType
 from lizard_progress.util import directories
 from lizard_progress.util import geo
@@ -72,16 +76,6 @@ logger = logging.getLogger(__name__)
 
 
 UiView  # Don't delete, it is imported by ``views/activity.py``
-
-
-# Decorator serving to assure the request is a AJAX request
-def ajax_request(view):
-    def wrapper(request, *args, **kwargs):
-        if request.is_ajax():
-            return view(request, *args, **kwargs)
-        else:
-            raise PermissionDenied
-    return wrapper
 
 
 class ProjectsMixin(object):